# infoschema: Don't truncate versions