--- a/infoschema/tables.go
+++ b/infoschema/tables.go
@@ -1770,7 +1770,7 @@ func FormatTiDBVersion(TiDBVersion string, isDefaultVersion bool) string {
 		if len(nodeVersion) > 0 && nodeVersion[0] == 'v' {
 			nodeVersion = nodeVersion[1:]
 		}
-		nodeVersions := strings.Split(nodeVersion, "-")
+		nodeVersions := strings.SplitN(nodeVersion, "-", 2)
 		if len(nodeVersions) == 1 {
 			version = nodeVersions[0]
 		} else if len(nodeVersions) >= 2 {