<!--

Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and 
linking the relevant issues via the "close" or "ref".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Issue Number: ref #16302 

Problem Summary:

Before (`7.3.0-alpha`):
```
mysql> TABLE information_schema.CLUSTER_INFO;
+------+-----------------+-----------------+-------------+------------------------------------------+---------------------+---------------+-----------+
| TYPE | INSTANCE        | STATUS_ADDRESS  | VERSION     | GIT_HASH                                 | START_TIME          | UPTIME        | SERVER_ID |
+------+-----------------+-----------------+-------------+------------------------------------------+---------------------+---------------+-----------+
| tidb | 127.0.0.1:4000  | 127.0.0.1:10080 | 7.3.0-alpha | 8c71149116f9439713604d3bf737156a2735b963 | 2023-07-06 15:14:07 | 12.962112656s |   1561139 |
| pd   | 127.0.0.1:2379  | 127.0.0.1:2379  | 7.2.0       | 0934e641129d8bac3c5975372b2ce492b81fd9db | 2023-07-06 15:13:54 | 25.962125837s |         0 |
| tikv | 127.0.0.1:20160 | 127.0.0.1:20180 | 7.2.0       | 12ce5540f9e8f781f14d3b3a58fb9442f03b6b29 | 2023-07-06 15:13:57 | 22.962131038s |         0 |
+------+-----------------+-----------------+-------------+------------------------------------------+---------------------+---------------+-----------+
3 rows in set (0.00 sec)
```

After (`7.3.0-alpha-128-g967699a509-dirty`):
```
mysql> TABLE information_schema.CLUSTER_INFO;
+------+-------------------+--------------------+-----------------------------------+------------------------------------------+---------------------------+--------------+-----------+
| TYPE | INSTANCE          | STATUS_ADDRESS     | VERSION                           | GIT_HASH                                 | START_TIME                | UPTIME       | SERVER_ID |
+------+-------------------+--------------------+-----------------------------------+------------------------------------------+---------------------------+--------------+-----------+
| tidb | 192.168.1.54:4000 | 192.168.1.54:10080 | 7.3.0-alpha-128-g967699a509-dirty | 967699a50938848bcd3504a43856d5487af2d0b3 | 2023-07-06T15:24:54+02:00 | 552.343537ms |         1 |
| tikv | store1            |                    |                                   |                                          |                           |              |         0 |
+------+-------------------+--------------------+-----------------------------------+------------------------------------------+---------------------------+--------------+-----------+
2 rows in set (0.00 sec)
```

### What is changed and how it works?

### Check List

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [ ] Integration test
- [ ] Manual test (add detailed scripts or steps below)
- [ ] No code

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [x] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- compatibility change, improvement, bugfix, and new feature need a release note -->

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

```release-note
The versions in `information_schema.cluster_info` are no longer truncated on the first `-`.
```